Move GPU startup helpers out of main
